[Security Solution] Consolidate data view manager test mocks behind one entry
Prep for extracting data_view_manager into a package by giving external tests a
single mocks entry point instead of deep-importing individual fixture files:

- Add `data_view_manager/mocks/index.ts` re-exporting the DataView fixtures,
  timeline data view specs, and the redux state mock; migrate all external test
  consumers to import from it.
- Export `mockTimelineDataViewId` from the redux state mock and use it in the
  timeline_save middleware test instead of reaching into `state.dataViewManager`.

`common/mock/global_state.ts` keeps importing the lean redux state mock directly
(it is imported by ~800 test files) to avoid pulling the DataView fixtures into
every test.
