+++ x-pack/solutions/security/plugins/security_solution/public/common/components/visualization_actions/lens_attributes/hosts/kpi_unique_ips_area.test.ts
@@ -11,7 +11,7 @@ import { wrapper } from '../../mocks';
 import { useLensAttributes } from '../../use_lens_attributes';
 
 import { getKpiUniqueIpsAreaLensAttributes } from './kpi_unique_ips_area';
-import { getMockDataViewWithMatchedIndices } from '../../../../../data_view_manager/mocks/mock_data_view';
+import { getMockDataViewWithMatchedIndices } from '../../../../../data_view_manager/mocks';
 import { useDataView } from '../../../../../data_view_manager/hooks/use_data_view';
 
 jest.mock('../../../../utils/route/use_route_spy', () => ({
