+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_table/alerts_sub_grouping.test.tsx
@@ -15,7 +15,7 @@ import { groupingSearchResponse } from './grouping_settings/mock';
 import { useKibana } from '../../../common/lib/kibana';
 import { createTelemetryServiceMock } from '../../../common/lib/telemetry/telemetry_service.mock';
 import { useDataView } from '../../../data_view_manager/hooks/use_data_view';
-import { getMockDataViewWithMatchedIndices } from '../../../data_view_manager/mocks/mock_data_view';
+import { getMockDataViewWithMatchedIndices } from '../../../data_view_manager/mocks';
 import type { DataView } from '@kbn/data-views-plugin/common';
 import { TableId } from '@kbn/securitysolution-data-table';
 import { defaultGroupStatsAggregations } from './grouping_settings';
