+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/tools/correlations/components/related_attacks.test.tsx
@@ -22,7 +22,7 @@ import {
 } from '../../../../shared/components/test_ids';
 import { usePaginatedAlerts } from '../hooks/use_paginated_alerts';
 import { useDataView } from '../../../../../data_view_manager/hooks/use_data_view';
-import { getMockDataViewWithMatchedIndices } from '../../../../../data_view_manager/mocks/mock_data_view';
+import { getMockDataViewWithMatchedIndices } from '../../../../../data_view_manager/mocks';
 import { useIsInSecurityApp } from '../../../../../common/hooks/is_in_security_app';
 import { useAlertsPrivileges } from '../../../../../detections/containers/detection_engine/alerts/use_alerts_privileges';
