+++ x-pack/solutions/security/plugins/security_solution/public/overview/pages/data_quality.test.tsx
@@ -16,10 +16,7 @@ import { DataQuality } from './data_quality';
 import { useKibana } from '../../common/lib/kibana';
 import { KibanaRenderContextProvider } from '@kbn/react-kibana-context-render';
 import { useDataView } from '../../data_view_manager/hooks/use_data_view';
-import {
-  getMockDataView,
-  getMockDataViewWithMatchedIndices,
-} from '../../data_view_manager/mocks/mock_data_view';
+import { getMockDataView, getMockDataViewWithMatchedIndices } from '../../data_view_manager/mocks';
 import {
   defaultImplementation,
   withIndices,
